Extract base class from IdeTestCase and IdeTheoryTestCase
