+++ src/Microsoft.VisualStudio.Extensibility.Testing.Xunit.Shared/Threading/IdeTheoryTestCase.cs
@@ -10,53 +10,17 @@ namespace Xunit.Threading
     using Xunit.Abstractions;
     using Xunit.Sdk;
 
-    public sealed class IdeTheoryTestCase : XunitTestCase
+    public sealed class IdeTheoryTestCase : IdeTestCaseBase
     {
         [EditorBrowsable(EditorBrowsableState.Never)]
         [Obsolete("Called by the deserializer; should only be called by deriving classes for deserialization purposes", error: true)]
-#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public IdeTheoryTestCase()
-#pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         {
         }
 
-        public IdeTheoryTestCase(IMessageSink diagnosticMessageSink, TestMethodDisplay defaultMethodDisplay, TestMethodDisplayOptions defaultMethodDisplayOptions, ITestMethod testMethod, VisualStudioVersion visualStudioVersion, object?[]? testMethodArguments = null)
-            : base(diagnosticMessageSink, defaultMethodDisplay, defaultMethodDisplayOptions, testMethod, testMethodArguments)
+        public IdeTheoryTestCase(IMessageSink diagnosticMessageSink, TestMethodDisplay defaultMethodDisplay, TestMethodDisplayOptions defaultMethodDisplayOptions, ITestMethod testMethod, VisualStudioVersion visualStudioVersion, string rootSuffix, object?[]? testMethodArguments = null)
+            : base(diagnosticMessageSink, defaultMethodDisplay, defaultMethodDisplayOptions, testMethod, visualStudioVersion, rootSuffix, testMethodArguments)
         {
-            SharedData = WpfTestSharedData.Instance;
-            VisualStudioVersion = visualStudioVersion;
-
-            if (!IdeTestCase.IsInstalled(visualStudioVersion))
-            {
-                SkipReason = $"{visualStudioVersion} is not installed";
-            }
-        }
-
-        public VisualStudioVersion VisualStudioVersion
-        {
-            get;
-            private set;
-        }
-
-        public new TestMethodDisplay DefaultMethodDisplay => base.DefaultMethodDisplay;
-
-        public new TestMethodDisplayOptions DefaultMethodDisplayOptions => base.DefaultMethodDisplayOptions;
-
-        public WpfTestSharedData SharedData
-        {
-            get;
-            private set;
-        }
-
-        protected override string GetDisplayName(IAttributeInfo factAttribute, string displayName)
-        {
-            var baseName = base.GetDisplayName(factAttribute, displayName);
-            return $"{baseName} ({VisualStudioVersion})";
-        }
-
-        protected override string GetUniqueID()
-        {
-            return $"{base.GetUniqueID()}_{VisualStudioVersion}";
         }
 
         public override Task<RunSummary> RunAsync(IMessageSink diagnosticMessageSink, IMessageBus messageBus, object[] constructorArguments, ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
@@ -73,21 +37,6 @@ namespace Xunit.Threading
             }
 
             return runner.RunAsync();
-        }
-
-        public override void Serialize(IXunitSerializationInfo data)
-        {
-            base.Serialize(data);
-            data.AddValue(nameof(VisualStudioVersion), (int)VisualStudioVersion);
-            data.AddValue(nameof(SkipReason), SkipReason);
-        }
-
-        public override void Deserialize(IXunitSerializationInfo data)
-        {
-            VisualStudioVersion = (VisualStudioVersion)data.GetValue<int>(nameof(VisualStudioVersion));
-            base.Deserialize(data);
-            SkipReason = data.GetValue<string>(nameof(SkipReason));
-            SharedData = WpfTestSharedData.Instance;
         }
     }
 }
